requirements.txt
fastapi==0.136.1
uvicorn[standard]==0.46.0
python-multipart==0.0.26
groq==1.2.0
anthropic==0.97.0
pypdf==6.10.2
python-dotenv==1.2.2
PyYAML==6.0.3
mcp==1.27.0
